game.c/main update
game.c has 3 games now, main has wireless implementation started, as
well as EEPROM
todo: ledMatrix. Also, timing is done.

+++ src/gpioPort.h
@@ -434,7 +434,7 @@ bool  gpio_config_alternate_function(uint32_t baseAddr, uint8_t pins);
 //    pins  -   A bit mask indicating which pins should be configured for the 
 //              specified port control register.
 //*****************************************************************************
-bool  gpio_config_port_control(uint32_t baseAddr, uint8_t pins);
+bool  gpio_config_port_control(uint32_t baseAddr, uint32_t pins);
 
 
 //************************************************************
